updated Recipe View
