+++ src/pages/RecipeView.module.scss
@@ -1,8 +1,8 @@
 .recipeHeader {
-    --border-width: 0px !important;
-    --background: transparent;
-    position: absolute;
-    top: 0rem;
+  --border-width: 0px !important;
+  --background: transparent;
+  position: absolute;
+  top: 0rem;
 }
 .headerImage {
   img {
@@ -11,68 +11,62 @@
     border-bottom: 5px solid var(--ion-color-main);
   }
 }
+.ingredientButtons {
+  ion-row {
+    display: flex;
+    align-items: flex-end;
+  }
+  ion-button {
+    margin: auto;
+  }
+}
+.headerInfo {
+  position: absolute;
+  top: 10rem;
+  z-index: 10;
 
-  .headerInfo {
-    position: absolute;
-    top: 10rem;
-    z-index: 10;
+  background-color: rgba($color: #000000, $alpha: 0.8);
+  color: white;
+  padding: 1rem;
+  border-top-right-radius: 10px;
+  border-bottom-right-radius: 10px;
 
-    background-color: rgba($color: #000000, $alpha: 0.8);
-    color: white;
-    padding: 1rem;
-    border-top-right-radius: 10px;
-    border-bottom-right-radius: 10px;
-
-    h1 {
-      font-size: 1.1rem;
-      padding: 0;
-      margin: 0;
-      font-family: "Calistoga";
-    }
-
-    p {
-      font-size: 0.9rem;
-      padding: 0;
-      margin: 0;
-      color: var(--ion-color-main);
-    }
+  h1 {
+    font-size: 1.1rem;
+    padding: 0;
+    margin: 0;
+    font-family: "Calistoga";
   }
 
+  p {
+    font-size: 0.9rem;
+    padding: 0;
+    margin: 0;
+    color: var(--ion-color-main);
+  }
+}
+.ingredientHeader:first-child {
+  font-size: 1.5rem;
+  text-align: center;
+}
 .ingredientHeader {
   width: 100%;
+  font-family: "Calistoga";
 
   ion-col {
     display: flex;
     flex-direction: column;
     justify-content: center;
-
-    ion-chip {
-      padding-top: 0%;
-      padding-bottom: 0%;
-      ion-col {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        text-align: center;
-      }
-      ion-col:nth-child(1) {
-        // flex-direction: column;
-        padding-left: 0;
-        border-right: #fff 1px solid;
-      }
-      ion-col:nth-child(2) {
-        padding-right: 0;
-
-        ion-icon {
-          font-size: 2rem;
-        }
-      }
-      small {
-        font-size: 65%;
-      }
+    ion-progress-bar {
+      height: 0.8rem;
+      border-radius: 5rem;
+    }
+    small {
+      font-size: 45%;
     }
   }
 }
+
 .itemLabel {
   h3 {
     display: flex;
@@ -91,7 +85,7 @@
 
   ion-button {
     height: 100%;
-    ion-icon{
+    ion-icon {
       color: maroon;
       font-size: 55px !important;
       --ionicon-stroke-width: 16px;
@@ -99,42 +93,40 @@
   }
 }
 .customBackBtn {
-    border-radius: 500px;
-    padding: 0rem;
-    background-color: rgba(0, 0, 0, 0.5);
-    border: 2px solid white;
-    display: flex;
-    flex-direction: row;
-    align-content: center;
-    align-items: center;
-    justify-content: center;
-  
-    ion-icon {
-      color: rgb(255, 255, 255);
-      height: 1.75rem;
-      width: 1.75rem;
-    }
+  border-radius: 500px;
+  padding: 0rem;
+  background-color: rgba(0, 0, 0, 0.5);
+  border: 2px solid white;
+  display: flex;
+  flex-direction: row;
+  align-content: center;
+  align-items: center;
+  justify-content: center;
+
+  ion-icon {
+    color: rgb(255, 255, 255);
+    height: 1.75rem;
+    width: 1.75rem;
   }
-  
-  .customFavoriteBtn {
-  
-    position: absolute;
-    top: 3rem;
-    right: 1.4rem;
-    border-radius: 500px;
-    padding: 0.5rem;
-    background-color: rgba(0, 0, 0, 0.5);
-    border: 2px solid white;
-    display: flex;
-    flex-direction: row;
-    align-content: center;
-    align-items: center;
-    justify-content: center;
-  
-    svg {
-  
-      color: rgb(255, 255, 255);
-      height: 1.75rem;
-      width: 1.75rem;
-    }
+}
+
+.customFavoriteBtn {
+  position: absolute;
+  top: 3rem;
+  right: 1.4rem;
+  border-radius: 500px;
+  padding: 0.5rem;
+  background-color: rgba(0, 0, 0, 0.5);
+  border: 2px solid white;
+  display: flex;
+  flex-direction: row;
+  align-content: center;
+  align-items: center;
+  justify-content: center;
+
+  svg {
+    color: rgb(255, 255, 255);
+    height: 1.75rem;
+    width: 1.75rem;
   }
+}
